For #100 - added drawing of polygons and circles to GeoFenceRenderer

+++ src/main/java/org/traccar/web/client/model/GeoFenceProperties.java
@@ -31,7 +31,7 @@ public interface GeoFenceProperties extends PropertyAccess<GeoFence> {
 
     ValueProvider<GeoFence, String> color();
 
-    ValueProvider<GeoFence, Double> radius();
+    ValueProvider<GeoFence, Float> radius();
 
     ValueProvider<GeoFence, GeoFenceType> type();
